+++ share/models/work_relations.py
@@ -7,7 +7,6 @@ from share.util import ModelGenerator
 class AbstractWorkRelation(ShareObject, metaclass=TypedShareObjectMeta):
     from_work = ShareForeignKey('AbstractCreativeWork', related_name='outgoing_%(class)ss')
     to_work = ShareForeignKey('AbstractCreativeWork', related_name='incoming_%(class)ss')
-    relation_type = TreeForeignKey('WorkRelationType')
 
     class Meta:
         unique_together = ('from_work', 'to_work', 'type')
